Move memory setting utils
