+++ src/base/Memory.nocuda.cc
@@ -1,26 +1,19 @@
-//----------------------------------*-C++-*----------------------------------//
+//---------------------------------*-C++-*-----------------------------------//
 // Copyright 2020 UT-Battelle, LLC, and other Celeritas developers.
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file Utils.i.hh
+//! \file Memory.cc
 //---------------------------------------------------------------------------//
+#include "Memory.hh"
 
 namespace celeritas
 {
-namespace detail
-{
 //---------------------------------------------------------------------------//
-/*!
- * Call memset on device data.
- */
-template<class T>
-void device_memset_zero(span<T> device_pointers)
+void device_memset(void*, int, size_type)
 {
-    device_memset(
-        device_pointers.data(), 0, device_pointers.size() * sizeof(T));
+    REQUIRE(0);
 }
 
 //---------------------------------------------------------------------------//
-} // namespace detail
 } // namespace celeritas
